requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
debugpy==1.5.1
celery==5.4.0
django-celery-beat
django-celery-results
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
pillow==10.4.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.0.0
django-modeltranslation==0.19.7
django-filter==24.3
gunicorn==20.1.0
boto3==1.34.140
django-storages==1.14.3
django-cors-headers==4.4.0
moviepy==1.0.3
google-auth==2.32.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
python-Levenshtein==0.26.0
